+++ apps/www-tron/next.config.mjs
@@ -6,6 +6,26 @@ const nextConfig = {
   images: {
     unoptimized: true,
   },
+
+  rewrites: async () => {
+    return [
+      {
+        source: '/',
+        destination: '/about',
+      },{
+      source: '/writing',
+        destination: '/blog'
+      },
+      {
+        source: '/sitemap.xml',
+        destination: 'https://claycurry.com/sitemap.xml',
+      },
+      {
+        source: '/robots.txt',
+        destination: 'https://claycurry.com/robots.txt',
+      },
+    ]
+  }
  
 }