+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 2026-01-19
+
+### Added
+- Resume page components imported from old portfolio
+- Tourney display font for headings
+
 ## 2026-01-18
 
 ### Added